pyproject.toml
[project]
name = "swarm-kit"
version = "0.1.0"
description = "A minimalist multi-agent orchestration library."
license = { text = "MIT" }
authors = [
    { name = "Moseleydev", email = "moseleydev7@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
]

[project.scripts]
swarm-kit = "swarm_kit.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"